feat: add CashAnnualizedCopy component for displaying annualized bonus copy in Cash section
